+++ SC-MMascotass/App.xaml
@@ -2,7 +2,7 @@
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:local="clr-namespace:SC_MMascotass"
-             StartupUri="Pages/Formularios/FormProveedores.xaml">
+             StartupUri="Pages/Formularios/FormCategoriasN.xaml">
    <Application.Resources>
          
     </Application.Resources>
