--- SC-MMascotass/Pages/Categorias.xaml
@@ -1,86 +1,0 @@
-﻿<UserControl x:Class="SC_MMascotass.Pages.Categorias"
-             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
-             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-             xmlns:local="clr-namespace:SC_MMascotass.Pages"
-             mc:Ignorable="d" 
-             d:DesignHeight="450" d:DesignWidth="800">
-    <Grid>
-        <Grid.ColumnDefinitions>
-            <ColumnDefinition/>
-        </Grid.ColumnDefinitions>
-        <Grid.RowDefinitions>
-            <RowDefinition Height="35"/>
-            <RowDefinition Height="75"/>
-            <RowDefinition/>
-        </Grid.RowDefinitions>
-        <StackPanel>
-            <TextBlock Text="Categorías"
-                        FontSize="24"
-                        />
-        </StackPanel>
-        <StackPanel Grid.Row="1" Orientation="Horizontal" HorizontalAlignment="Center" >
-            <Button     Background="Transparent"
-                        Foreground="Black"
-                        FontSize="14"
-                        BorderThickness="0"
-                        Name="btnNuevo"
-                        Padding="10" Click="btnNuevo_Click" Height="75" VerticalAlignment="Top">
-                <StackPanel Orientation="Vertical" >
-                    <Image Source="/Imagenes/iconos/nuevo.png" HorizontalAlignment="Center" Width="25" Height="25"/>
-                    <TextBlock Text="Nuevo" HorizontalAlignment="Center"/>
-
-                </StackPanel>
-
-            </Button>
-            <Button     Background="Transparent"
-                        Foreground="Black"
-                        FontSize="14"
-                        BorderThickness="0"
-                        Name="btnEditar"
-                        Padding="10" Click="btnEditar_Click">
-                <StackPanel Orientation="Vertical" >
-                    <Image Source="/Imagenes/iconos/editar.png" HorizontalAlignment="Center" Width="25" Height="25"/>
-                    <TextBlock Text="Editar" HorizontalAlignment="Center"/>
-
-                </StackPanel>
-
-            </Button>
-            <Button     Background="Transparent"
-                        Foreground="Black"
-                        FontSize="14"
-                        BorderThickness="0"
-                        Name="btnEliminar"
-                        Padding="10" Click="btnEliminar_Click">
-                <StackPanel Orientation="Vertical" >
-                    <Image Source="/Imagenes/iconos/eliminar.png" HorizontalAlignment="Center" Width="25" Height="25"/>
-                    <TextBlock Text="Eliminar" HorizontalAlignment="Center"/>
-
-                </StackPanel>
-
-            </Button>
-            <Button     Background="Transparent"
-                        Foreground="Black"
-                        FontSize="14"
-                        BorderThickness="0"
-                        Name="btnRefrescar"
-                        Padding="10" Click="btnRefrescar_Click">
-                <StackPanel Orientation="Vertical" >
-                    <Image Source="/Imagenes/iconos/recargar.png" HorizontalAlignment="Center" Width="25" Height="25"/>
-                    <TextBlock Text="Refrescar" HorizontalAlignment="Center"/>
-
-                </StackPanel>
-
-            </Button>
-        </StackPanel>
-        <StackPanel Grid.Row="2">
-            <DataGrid Name="dgClientes" AutoGenerateColumns="False" IsReadOnly="True" ItemsSource="{Binding Source=categorias}" MaxHeight="310" ColumnWidth="*">
-                <DataGrid.Columns>
-                    <DataGridTextColumn Header="ID" Binding="{Binding Id}" Width="50"/>
-                    <DataGridTextColumn Header="Nombre de la Categoría" Binding="{Binding NombreCategoria}" />
-                </DataGrid.Columns>
-            </DataGrid>
-        </StackPanel>
-    </Grid>
-</UserControl>
