--- SC-MMascotass/Pages/Formularios/FormCategorias.xaml
@@ -1,109 +1,0 @@
-﻿<Window x:Class="SC_MMascotass.Pages.Formularios.FormCategorias"
-        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:SC_MMascotass.Pages.Formularios" WindowStartupLocation="CenterScreen"
-        mc:Ignorable="d"
-        Title="Categoria" Height="216.394" Width="518.443">
-    <Grid Margin="10">
-        <Grid.ColumnDefinitions>
-            <ColumnDefinition/>
-        </Grid.ColumnDefinitions>
-        <Grid.RowDefinitions>
-            <RowDefinition Height="30"/>
-            <RowDefinition Height="75"/>
-            <RowDefinition/>
-        </Grid.RowDefinitions>
-
-        <StackPanel Orientation="Horizontal" Visibility="Visible" Name="spNuevaCategoria">
-            <Image Source="/Imagenes/categoria.png" Width="25" Margin="0 0 10 0"></Image>
-            <TextBlock Text="Nueva Categoría"
-                        FontSize="18"
-                       VerticalAlignment="Center"
-                        />
-        </StackPanel>
-        <StackPanel Grid.Row="1" Orientation="Horizontal" HorizontalAlignment="Center" Visibility="Visible" Name="spButton1">
-            <Button     Background="Transparent"
-                        Foreground="Black"
-                        FontSize="14"
-                        BorderThickness="0"
-                        Name="btnGuardar"
-                        Padding="10" Click="btnGuardar_Click">
-                <StackPanel Orientation="Vertical" >
-                    <Image Source="/Imagenes/iconos/guardar.png" HorizontalAlignment="Center" Width="25" Height="25"/>
-                    <TextBlock Text="Guardar" HorizontalAlignment="Center"/>
-
-                </StackPanel>
-
-            </Button>
-            <Button     Background="Transparent"
-                        Foreground="Black"
-                        FontSize="14"
-                        BorderThickness="0"
-                        Name="btnRestablecer"
-                        Padding="10" Click="btnRestablecer_Click">
-                <StackPanel Orientation="Vertical" >
-                    <Image Source="/Imagenes/iconos/recargar.png" HorizontalAlignment="Center" Width="25" Height="25"/>
-                    <TextBlock Text="Limpiar" HorizontalAlignment="Center"/>
-
-                </StackPanel>
-
-            </Button>
-        </StackPanel>
-
-        <StackPanel Orientation="Horizontal" Visibility="Hidden" Name="spEditarCategoria">
-            <Image Source="/Imagenes/categoria.png" Width="25" Margin="0 0 10 0"></Image>
-            <TextBlock Text="Editar Categoría"
-                       FontSize="18"
-                       VerticalAlignment="Center">
-            </TextBlock>
-        </StackPanel>
-        <StackPanel Grid.Row="1" Orientation="Horizontal" HorizontalAlignment="Center" Visibility="Hidden" Name="spButton2">
-            <Button     Background="Transparent"
-                        Foreground="Black"
-                        FontSize="14"
-                        BorderThickness="0"
-                        Name="bntAceptar"
-                        Padding="10" Click="bntAceptar_Click">
-                <StackPanel Orientation="Vertical" >
-                    <Image Source="/Imagenes/iconos/guardar.png" HorizontalAlignment="Center" Width="25" Height="25"/>
-                    <TextBlock Text="Aceptar" HorizontalAlignment="Center"/>
-
-                </StackPanel>
-
-            </Button>
-            <Button     Background="Transparent"
-                        Foreground="Black"
-                        FontSize="14"
-                        BorderThickness="0"
-                        Name="btnCancelar"
-                        Padding="10" Click="btnCancelar_Click">
-                <StackPanel Orientation="Vertical" >
-                    <Image Source="/Imagenes/iconos/recargar.png" HorizontalAlignment="Center" Width="25" Height="25"/>
-                    <TextBlock Text="Cancelar" HorizontalAlignment="Center"/>
-
-                </StackPanel>
-
-            </Button>
-        </StackPanel>
-        <StackPanel Grid.Row="2">
-            <Grid>
-                <Grid.ColumnDefinitions>
-                    <ColumnDefinition Width="*"/>
-                    <ColumnDefinition Width="*"/>
-                </Grid.ColumnDefinitions>
-                <Grid.RowDefinitions>
-                    <RowDefinition Height="auto"/>
-                </Grid.RowDefinitions>
-
-
-
-                <Label Content="Categoría" Grid.Row="0" FontSize="15" VerticalAlignment="Center" Margin="0 4 0 0"/>
-
-                <TextBox Name="txtCategoria" FontSize="15" Grid.Column="1" Grid.Row="0" Margin="0 4 0 0"/>
-
-            </Grid>
-        </StackPanel>
-    </Grid>
-</Window>
