+++ SC-MMascotass/Pages/Formularios/FormInventario.xaml
@@ -98,7 +98,7 @@
 
                 <!--<StackPanel Orientation="Horizontal" Grid.Column="3"  HorizontalAlignment="Right">-->
                 <TextBox Margin="0 4 0 0" Padding="5, 3, 5, 3"  Name="txtAuCategoria" KeyUp="txtAuCategoria_KeyUp" Grid.Column="3"/>
-                <Button Width="50" Margin="0 4 0 0" ToolTip="Nuevo Producto" Name="btnNuevoCategoria" Background="#FCA311" Click="btnNuevoCliente_Click" Grid.Column="3" HorizontalAlignment="Right">
+                <Button Width="50" Margin="206.6,4,0,0.4" ToolTip="Nueva Categoria" Name="btnNuevoCategoria" Background="#FCA311" Click="btnNuevoCliente_Click" Grid.Column="3" HorizontalAlignment="Left">
                     <Image Source="/Imagenes/iconos/nuevo.png" Width="20"/>
                 </Button>
                 <!--</StackPanel>-->
