+++ SC-MMascotass/Pages/Formularios/FormInventario.xaml.cs
@@ -140,7 +140,7 @@ namespace SC_MMascotass.Pages.Formularios
         {
             bool found = false;
             var border = (autoCompleteCategorias.Parent as ScrollViewer).Parent as Border;
-            var data = Constructores.Procedimientos.MonstrarCategorias22();
+            var data = Constructores.Procedimientos.MostrarCategorias22();
 
             string query = (sender as TextBox).Text;
 
@@ -213,7 +213,7 @@ namespace SC_MMascotass.Pages.Formularios
         private void btnNuevoCliente_Click(object sender, RoutedEventArgs e)
         {
             // Mostrar el formulario de ingreso de categorias
-            FormCategorias categorias = new FormCategorias(false);
+            Formularios.FormCategoriasN categorias = new Formularios.FormCategoriasN();
             categorias.Show();
         }
