+++ SC-MMascotass/SC-MMascotass.csproj
@@ -105,17 +105,14 @@
     <Compile Include="Constructores\Validaciones.cs" />
     <Compile Include="database\Conexion.cs" />
     <Compile Include="Constructores\ErrorLog.cs" />
-    <Compile Include="Pages\Categorias.xaml.cs">
-      <DependentUpon>Categorias.xaml</DependentUpon>
-    </Compile>
     <Compile Include="Constructores\HistorialVacunacion.cs" />
     <Compile Include="Constructores\InventarioC.cs" />
     <Compile Include="Constructores\Mascota.cs" />
     <Compile Include="Pages\Expedientes.xaml.cs">
       <DependentUpon>Expedientes.xaml</DependentUpon>
     </Compile>
-    <Compile Include="Pages\Formularios\FormCategorias.xaml.cs">
-      <DependentUpon>FormCategorias.xaml</DependentUpon>
+    <Compile Include="Pages\Formularios\FormCategoriasN.xaml.cs">
+      <DependentUpon>FormCategoriasN.xaml</DependentUpon>
     </Compile>
     <Compile Include="Pages\Formularios\FormCliente.xaml.cs">
       <DependentUpon>FormCliente.xaml</DependentUpon>
@@ -242,15 +239,11 @@
     <Compile Include="SqlServerTypes\Loader.cs" />
     <Compile Include="Constructores\Usuario.cs" />
     <Compile Include="Constructores\vacunaciones.cs" />
-    <Page Include="Pages\Categorias.xaml">
-      <SubType>Designer</SubType>
-      <Generator>MSBuild:Compile</Generator>
-    </Page>
     <Page Include="Pages\Expedientes.xaml">
       <SubType>Designer</SubType>
       <Generator>MSBuild:Compile</Generator>
     </Page>
-    <Page Include="Pages\Formularios\FormCategorias.xaml">
+    <Page Include="Pages\Formularios\FormCategoriasN.xaml">
       <SubType>Designer</SubType>
       <Generator>MSBuild:Compile</Generator>
     </Page>
